feat: implement TotalNavPanel component with month selection and chart data handling
- Added TotalNavPanel component to display total NAV and allow month selection.
- Integrated useChartData hook for fetching and managing chart data from socket and API.
- Removed old TotalNavChart component and replaced it with a more modular approach.
- Updated types to include TNavChartDataFromApi for better data handling.
- Adjusted imports in Home index to reflect new component structure.
