+++ src/pages/Home/TotalNavPanel/components/TotalNavPanel.tsx
@@ -106,24 +106,6 @@ export default function TotalNavPanel() {
         </div>
       </div>
 
-      {/* Optional: Error display */}
-      {/* {navError && (
-        <div className="px-6 pb-4">
-          <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded">
-            <span className="block sm:inline">{navError}</span>
-            <button
-              onClick={() => {
-                // You could add reconnect logic here if needed
-                console.log("Reconnecting...");
-              }}
-              className="float-right font-bold text-red-500 hover:text-red-800"
-            >
-              Retry
-            </button>
-          </div>
-        </div>
-      )} */}
-
       {/* Chart */}
       <div className="pr-6">
         <TotalNavChart
